Internally catch FNFE in FileConfig

+++ src/main/java/de/mlessmann/confort/config/FileConfig.java
@@ -1,5 +1,6 @@
 package de.mlessmann.confort.config;
 
+import de.mlessmann.confort.api.except.ParseException;
 import de.mlessmann.confort.api.lang.IConfigLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,11 +24,26 @@ public class FileConfig extends Config {
         if (directory == null || (!directory.isDirectory() && !directory.mkdirs())) {
             throw new IOException("Cannot create configuration parent directory: " + directory);
         }
+
+        if (!file.isFile() && !file.createNewFile()) {
+            throw new IOException(String.format("Cannot create configuration file: %s", file.getPath()));
+        }
+
         return new OutputStreamWriter(new FileOutputStream(file), getEncoding());
     }
 
     @Override
     protected Reader produceReader() throws IOException {
         return new InputStreamReader(new FileInputStream(file), getEncoding());
+    }
+
+    @Override
+    public synchronized void load() throws IOException, ParseException {
+        try {
+            super.load();
+        } catch (FileNotFoundException e) {
+            createRoot();
+            save();
+        }
     }
 }
